Add support for AWS ALB
This adds support for the AWS Application Load Balancer

ALB doesn't support custom cipher lists
* "Application Load Balancers do not support custom security policies"
* https://docs.aws.amazon.com/elasticloadbalancing/latest/application/create-https-listener.html#describe-ssl-policies
As a result this maps our Old and Intermediate profiles to the closest similar
AWS policy.
* Old : ELBSecurityPolicy-TLS-1-0-2015-04
* Intermediate : ELBSecurityPolicy-TLS-1-2-2017-01
* Modern : unsupported

This CloudFormation template also contains a bunch of example infrastructure
which is needed to provision an ALB. I've added a comment to clarify that
for most people they'll be able to ignore all of it as they'll likely have
their own existing infrastructure.

The template includes an HTTP to HTTPS redirect, though HSTS isn't supported
at the ALB level (it could be done by the backend server though)

The server version comes from the 2018-07-25 revision to AWS ALB that added
support for redirects which we use to redirect HTTP to HTTPS
* https://docs.aws.amazon.com/elasticloadbalancing/latest/application/doc-history.html