Adds support for Traefik v2
This commit adds support for the new Traefik v2 configuration format. Starting
with Traefik 2.0, there are now two parts of a Traefik configuration, static
and dynamic configuration. This means the configuration format changed
considerably from a relatively simple format to something more complex.

In `config.js`, updated the latest version of Traefik. In `traefik.hbs`, added
the complete template for the new configuration format, as found in the Traefik
documentation. Especially relevant are the following two documentation
sections:

- [ TLS Configuration Is Now Dynamic, per Router ]
  https://docs.traefik.io/migration/v1-to-v2/#tls-configuration-is-now-dynamic-per-router

- [ HTTP to HTTPS Redirection Is Now Configured on Routers ]
  https://docs.traefik.io/migration/v1-to-v2/#http-to-https-redirection-is-now-configured-on-routers

Fixes #66