golang: remove note about ECDSA and compatibility
It looks like this was added because Go does not support DHE, but that
has no connection I am aware of to the type of certificate. Clients can
and do use ECDHE with RSA certificates through the ECDHE-RSA cipher
suites. If anything, there are clients that don't support ECDSA
certificates, so RSA certificates will have wider compatibility.