Update Go generator (#101)
* golang: the name of the language is Go

golang is just an abbreviation for "Go language" and should never be
used capitalized.

* golang: fix minimum TLS 1.3 version

In Go 1.12, TLS 1.3 was available but required opting in with an
environment variable that the config does not set. Go 1.12 is out of
security support anyway.

* golang: omit default settings

Clean up the config by not printing empty or false (which are the
default zero value in Go) Config field values.

* golang: do not disable HTTP/2

Unclear if this was intentional, but setting TLSNextProto to an empty
map overrides the default one which serves HTTP/2 for "h2".

* golang: remove note about ECDSA and compatibility

It looks like this was added because Go does not support DHE, but that
has no connection I am aware of to the type of certificate. Clients can
and do use ECDHE with RSA certificates through the ECDHE-RSA cipher
suites. If anything, there are clients that don't support ECDSA
certificates, so RSA certificates will have wider compatibility.

* golang: recommend setting Server timeouts

This is not a TLS configuration, but since this config is likely to get
exposed to the Internet and it shows how to create a Server, it should
inform the developer of the importance of setting timeouts to avoid
SlowLoris attacks.

* golang: add missing comma

The current config doesn't compile because Go needs a trailing comma
before a newline.

* golang: add redirect HTTP to HTTPS