Update to use Webpack 5
With this change, the static assets like fonts are hosted at the root of the path
instead of within their subdirectories like `/fonts`

I've been unable to figure out how to get this to work correctly where
the SASS `url` references to fonts would point to `/fonts`. To work around
this, I've commented out the `outputPath` in the `asset/resource` copying over
of the fonts.